search/app/requirements.txt
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
flasgger==0.9.5
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
mistune==2.0.4
pyrsistent==0.18.1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
